Fix asset metadata files and gitignore
Overwriting the recently-generated asset metadata files with the ones from the original project, since we want to preserve those links.
They were lost because there was an entry for "*.meta", so they were not uploaded to the repository when uploading a 'fresh' copy through the updated gitignore filter.
I also took that rule out so it won't break in the future.