Comments and Organization.
Added copyright notice to all scripts; please add your name as a
contributor to pages you work on. So far the code was fairly clean, so I
just tidied things up a tad. Feel free to use a few more comments when
variables aren't glaringly obvious. Using  [Tooltip("message")] allows
the user to see the comment message in the Editor when hovering over
public variables, so feel free to make use of that more as well.