Noteable Changes: -Lobby Screen removed and merged into "host" scene -Using potions, attacking, or healing now immediately ends a turn -netcode and other redundant code removed to best of my ability. Traces still remain - code is generally more readable -added a basic "end game" screen with a button which returns to main menu -fixed issues where dead enemies still had turns - game should design as intended now
Noteable issues:
-Cleric AI tends to attempt to attack enemies
-Grenadier makes questionably smart targetting decisions
-Bug i cannot work out: sometimes, the game returns an error when attempting to initialize end turn but works perfectly a second later - it has something to do with dequeing the next unit. It does provide a console error for backtracing though.

NOTES:
MATCHES CAN BE MADE SOLEY OUT OF AI. may be useful for monitoring AI behavior.
If you find netcode that you can not work its use, theres a decent chance that it can be removed with minimal dependency issues.
Due to the way damage dealing is handled, the game automatically ends the turn when it detects that any character has taken damage. Not an issue immediately; but might conflict with other changes if more functionality is added down the road.